Duplicated code refactored
